GROK-13043: Bioreactor: Remove FAE features
Initial values, labels, parameters and other FAE features are changed.
