+++ packages/Bioreactors/README.md
@@ -1,3 +1,3 @@
 # Bioreactors
 
-Bioreactors is a [package](https://datagrok.ai/help/develop/develop#packages) for the [Datagrok](https://datagrok.ai) platform. Simulation of bioreactors are provided.
+Bioreactors is a [package](https://datagrok.ai/help/develop/develop#packages) for the [Datagrok](https://datagrok.ai) platform. Mechanism of Controlled Fab-Arm Exchange is simulated (see paper Dennis R. Goulet et. al. "Kinetic mechanism of controlled Fab-arm exchange for the formation of bispecific immunoglobulin G1 antibodies", DOI: https://doi.org/10.1074/jbc.ra117.000303).
